- Fixing a couple compile warnings  - Fixing a bug where Sacrifice would explode your character if you used it on the air  - Fixing a bug where Sacrifice would kill you (it should leave you at 1 HP)  - Fixing a bug where Hyper Body wouldn't be reset on death so you could spawn with 160% of your base MP
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/branches/buis_vana_overhaul@3262 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/LuaScriptable.cpp
@@ -1566,7 +1566,7 @@ int LuaExports::getMonth(lua_State *luaVm) {
 }
 
 int LuaExports::getNearestMinute(lua_State *luaVm) {
-	lua_pushinteger(luaVm, TimeUtilities::getDistanceToNextMinuteMark(lua_tointeger(luaVm, 1)).count());
+	lua_pushinteger(luaVm, static_cast<int32_t>(TimeUtilities::getDistanceToNextMinuteMark(lua_tointeger(luaVm, 1)).count()));
 	return 1;
 }
